일시적 출력 파일 자동 삭제 로직 추가
`delete_outputs` 함수를 통해 일시적으로 생성된 출력 파일을 자동으로 삭제하도록 구현했습니다. 이로 인해 불필요한 파일 누적을 방지하고 디스크 공간을 절약할 수 있습니다. feature/cancel_in_queue
